+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# upcoming
+# Upcoming
 
 * BUGFIX: add a watch in radio-input to assign model value only when it's established
+* Add missing module dependencies for vr.slider and angular-bind-html-compile
+* Add style classes for grouped-select
+* Add angular-bootstrap bower dependency
 
 # 6.0.1